Add remote prefix option
This allows a built site to be committed into a subdirectory on GitHub
Pages, and can be used to deploy multiple versions of documentation
alongside each other (for example a dev and prod deployment).

Note: because `ghp_import` cannot keep existing files in a directory,
it is not possible to a deploy a site with no prefix (i.e. into the root
directory) as well as a site with a prefix (e.g. `dev`). The un-prefixed
site will remove the prefixed site from the branch when it is committed.

Signed-off-by: Ben Agricola <ben+git@agrico.la>